TYPING-TEXT-EFFECT
The Typing Animation project uses HTML, CSS, and JavaScript to create a cool typing effect on your website. It adds a dynamic touch, making text appear as if it's being typed in real-time. It's a fun and engaging way to showcase information, giving your website a stylish and interactive vibe.